single MessageService file
